extract seed database to separate file to make ad hoc use easier
